block queue info
