+++ ethcore/src/client/client.rs.in
@@ -14,10 +14,32 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 
+use std::marker::PhantomData;
+use std::path::PathBuf;
+use std::collections::{HashSet, HashMap};
+use std::ops::Deref;
+use std::mem;
+use std::collections::VecDeque;
+use std::sync::*;
+use std::path::Path;
+use std::fmt;
+use util::Itertools;
+
+// util
+use util::numbers::*;
+use util::panics::*;
+use util::network::*;
+use util::io::*;
+use util::rlp;
+use util::sha3::*;
+use util::{UtilError, CryptoError, Bytes, Signature, Secret, ec};
+use util::rlp::{encode, decode, RlpStream, Rlp};
+use util::journaldb;
+use util::journaldb::JournalDB;
 use std::path::PathBuf;
 use std::sync::atomic::{AtomicUsize, Ordering as AtomicOrdering};
-use util::*;
-use util::panics::*;
+
+// other
 use views::BlockView;
 use error::{Error, ImportError, ExecutionError, BlockError, ImportResult};
 use header::{BlockNumber, Header};
@@ -478,7 +500,7 @@ impl Client {
 	}
 }
 
-//#[derive(Ipc)]
+#[derive(Ipc)]
 impl BlockChainClient for Client {
 	fn call(&self, t: &SignedTransaction, analytics: CallAnalytics) -> Result<Executed, ExecutionError> {
 		let header = self.block_header(BlockID::Latest).unwrap();
@@ -770,7 +792,7 @@ impl BlockChainClient for Client {
 		self.build_last_hashes(self.chain.best_block_hash())
 	}
 
-	fn import_transactions(&self, transactions: Vec<SignedTransaction>) -> Vec<Result<TransactionImportResult, Error>> {
+	fn import_transactions(&self, transactions: Vec<SignedTransaction>) -> Vec<Result<TransactionImportResult, String>> {
 		let fetch_account = |a: &Address| AccountDetails {
 			nonce: self.latest_nonce(a),
 			balance: self.latest_balance(a),
