+++ ethcore/src/miner/transaction_queue.rs
@@ -197,6 +197,7 @@ struct VerifiedTransaction {
 	/// transaction origin
 	origin: TransactionOrigin,
 }
+
 impl VerifiedTransaction {
 	fn new(transaction: SignedTransaction, origin: TransactionOrigin) -> Result<Self, Error> {
 		try!(transaction.sender());
